# Tessellate: Implement precomputed vertices for rounding